refactored related products to 3 separate files
